pyproject.toml
[project]
name = "buena-context"
version = "0.1.0"
description = "Living building memory — context compiler for property management"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.136.1",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "structlog>=24.4",
    "httpx>=0.28",
    "anthropic>=0.97.0",
    "duckdb>=1.5.2",
    "pypdf>=6.10.2",
    "fastmcp>=3.2.4",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "anyio>=4.7",
    "ruff>=0.8",
    "ty>=0.0.1a1",
]

[build-system]
requires = ["uv_build>=0.11,<0.12"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "app"
module-root = ""

# ---- Ruff (lint + format) ----
[tool.ruff]
target-version = "py313"
line-length = 100
exclude = [".agents"]

[tool.ruff.lint]
select = [
    "E", "F", "W",      # pyflakes/pycodestyle
    "I",                # isort
    "B",                # bugbear
    "UP",               # pyupgrade
    "ASYNC",            # async best practices
    "S",                # bandit security
    "SIM",              # simplify
    "RUF",              # ruff-specific
    "PL",               # pylint subset
    "TID",              # tidy imports
    "PT",               # pytest style
    "ANN",              # type annotations
    "FAST",             # fastapi-specific
]
ignore = [
    "ANN401",           # allow Any in narrow cases
    "PLR0913",          # many args ok in routes
    "S101",             # assert ok in tests
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S", "ANN", "PLR2004"]

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true

# ---- ty (type checker) ----
[tool.ty.src]
include = ["app", "tests"]

# ---- pytest ----
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers --strict-config"
